#5 refactor structs (#6)
* Refactor structure

We now use Matchthread -> Comments -> Streams. This helps us organise the data in a better way since it works more like the reddit scraping. The web functionality is just duct-taped together right now to work with the new structure and needs to be rebuilt in some parts.

* Fix web still using the Streams array in places

Added NumStreams temporarily to fix the behavior and all relevant checks in the templates

* Restrict drone tasks to push event

* Redesign web views to fit new structure